requirements.txt
web3==6.11.1
Flask==3.0.0
pycryptodome==3.19.0
pytest==7.4.3
